pyproject.toml
# pyproject.toml
[project]
name = "truck-marketplace-bot"
version = "0.1.0"
description = "Telegram бот для биржи грузоперевозок и спецтехники"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiogram>=3.10",
    "aiofiles>=23.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "python-dotenv>=1.0",
    "aiosqlite>=0.19",  # Для async SQLite
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "isort>=5.12",
    "mypy>=1.7",
]
docker = [
    "asyncpg>=0.29",  # Для PostgreSQL в Docker
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true